refactor(review): keep context contract within size guard
